refactor(project): :recycle: move source code into a subpackage
