+++ gemax/internal/multierr/mulitierr.go
