+++ gemax/internal/tester/doc.go
