+++ gemax/server_utils_test.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/ninedraft/gemax"
-	"github.com/ninedraft/gemax/status"
+	"github.com/ninedraft/gemax/gemax"
+	"github.com/ninedraft/gemax/gemax/status"
 )
 
 func TestServeContent(test *testing.T) {
